refactor(core): migrate remote connect bot support owners
