+++ src/crates/services/services-integrations/Cargo.toml
@@ -101,6 +101,7 @@ remote-connect = [
     "base64",
     "bitfun-runtime-ports",
     "chrono",
+    "dirs",
     "futures",
     "hostname",
     "image",
